basic functionality working (sorta)
